Improved type safety and code quality across modules:   - Added type annotations throughout codebase (Dict, Optional, explicit types)   - Fixed multi-line condition formatting with proper parentheses   - Refined error handling with string conversions and None checks   - Updated MLflow security warnings with detailed CVE information   - Simplified import statements by removing unnecessary parentheses   - Added documentation files for ansible, api keys, csharp, and vscode integrations   - Configured mypy for type checking

+++ modules/suggestion/feedback.py
@@ -437,7 +437,7 @@ class FeedbackManager:
         if not ratings:
             return None
 
-        return sum(ratings) / len(ratings)
+        return float(sum(ratings) / len(ratings))
 
     def get_approval_count(self, suggestion_id: str) -> int:
         """Get the number of approvals for a suggestion.
